updates on the aurein MSMs
Aurein 1.2 - 250ps Tuning was 👍🏻 from last time. In this notebook, I just added the aligned (again) structures. Each of the sample structures from each state was separated from the bundle, visualized, and aligned using VMD (noH alignment and backbone alignment). Then they were made pretty in Chimera. 

Aurein 2.2 - This is the 100ps Tuning. It was done in hopes of a better free energy surface. There are improvements but I don't think we'll get a completely continuous free energy surface even if we go with 50ps or fewer frames. A comparison plot of 100ps vs. 250ps tuning is added as well.

Aurein 2.5 - This is the 250ps Tuning as the last time. In this notebook, I chose a shorter lag time of 2ns and two states.

Aurein 2.5_alp - 250ps Tuning. Looks odd! The ITS plot doesn't converge.

Aurein 3.1 - 250ps Tuning. The free energy surface is different in comparison to other peptides but nothing odd really.

Aurein 4.1 - 100ps and 250ps Tuning. The VAMP2 score in just increasing. Although the slope becomes smaller for the larger lag times.

Aurein 5.1 - 100ps and 250ps Tuning. Same as the Aurein 4.1.